Remove unused log and OS imports across multiple files
This commit eliminates the unused "log" import from the `redis.go` and `event.go` files in the pkg directory. It also removes "os", "os/signal", and "syscall" imports from the `component_test.go` file in tests. This refactoring is necessary to clean up the codebase and improve its readability. The redundant use of `CorrelationPublisherDecorator` was also removed as it is no longer needed.