Added functionality for handling ticket limits
This commit implements the feature that keeps track of ticket limits for shows and handles different scenarios such as booking tickets and overbooking. Necessary HTTP calls for testing the functionality were added in the testing files along with the related struct type definitions. Changes were made to the SQL transactions in the Postgres Repository to accommodate the aforementioned features.