Refactored 'Service' to 'App' in project files
Refactored the 'service' package to 'app' for readability and consistency across all project files. Changes include renaming the package and all use instances of 'Service', including constructor functions and method receiver names. This is to align our application architecture with Go standard project layout and for clarity on the application's main function.