Refactor pubsub package structure for better organization
Transferred all commands and events from "pubsub/command" and "pubsub/event" to "pubsub/handlers/command" and "pubsub/handlers/event" respectively. This restructuring is important to improve code organization and maintainability. The functionality remains unaffected as the change only involves relocating files and not modifying the actual codes. Consequently, the changes in imports seen in the diff are due to the updated file locations.